requirements.txt
joblib
matplotlib
numpy
streamlit
pyyaml
